jobs/pdf-report/requirements.txt
boto3
requests
matplotlib
fpdf2
pandas